backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
transformers==4.35.2
torch==2.1.1
torchvision==0.16.1
pillow==10.1.0
whisper-openai==20231117
indic-parler-tts==0.0.1
numpy==1.24.3
python-multipart==0.0.6
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0
scikit-learn==1.3.2
opencv-python==4.8.1.78
requests==2.31.0
